pyproject.toml
[project]
name = "new-admin-panel-sprint-3"
version = "0.3.0"
description = "Мой проект на курсе Яндекс Практикум"
authors = [
    {name = "clasterpack",email = "clasterpack@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "psycopg (>=3.2.4,<4.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uwsgi (>=2.0.28,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "elasticsearch (>=8.17.2,<9.0.0)",
    "backoff (>=2.2.1,<3.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
wemake-python-styleguide = "^1.0.0"
isort = "^6.0.0"
flake8 = "^7.1.2"

